Updated Tags, Collectible items, Attacking state
Added tags for each individual collectible item, as well as added the
ability to 'pick up' those items though none do anything at this point
of time. Additionally added an attacking state to the PlayerControl  in
which the game knows when the player is and is not attacking. Lastly
fixed the bug where if the player held a move key while holding the
attack key, the character would move/attack indefinitely.